requirements.txt
swapi>=0.1.3
requests
numpy>=1.20.2
pandas>=1.2.4
Faker>=8.1.2
Django>=3.0,<4.0
psycopg2-binary>=2.8
